+++ src/Phan/Analysis/ReturnTypesAnalyzer.php
@@ -6,10 +6,12 @@ use Phan\Config;
 use Phan\Issue;
 use Phan\IssueFixSuggester;
 use Phan\Language\Element\FunctionInterface;
+use Phan\Language\Element\Func;
 use Phan\Language\Element\Method;
 use Phan\Language\FQSEN\FullyQualifiedClassName;
 use Phan\Language\Type\GenericArrayType;
 use Phan\Language\Type\TemplateType;
+use Phan\Language\Type\VoidType;
 use Phan\Language\UnionType;
 
 class ReturnTypesAnalyzer
@@ -119,6 +121,11 @@ class ReturnTypesAnalyzer
                         }
                     }
                 }
+            }
+        }
+        if ($return_type->isEmpty() && !$method->getHasReturn()) {
+            if ($method instanceof Func || ($method instanceof Method && $method->isPrivate())) {
+                $method->setUnionType(VoidType::instance(false)->asUnionType());
             }
         }
     }